improve memory consumption
- The id of a CoordNode/RouteNode is an int, not a long.

- Don't save incoming and outgoing arcs in RouteNode. We 
always create the reverse arc, so we can simply add a pointer
to the reverse arc in each RouteArc instance. 
This reduces the number of ArrayList instances.

This change saves around 10 MB heap for a tile with ~110000 RouteNodes.


git-svn-id: http://svn.mkgmap.org.uk/mkgmap/branches/performance2@3257 25d90789-57f7-4ee0-8453-03a3dfeeeb22

+++ src/uk/me/parabola/imgfmt/app/Coord.java
@@ -124,7 +124,10 @@ public class Coord implements Comparable<Coord> {
 		return longitude;
 	}
 
-	public long getId() {
+	/**
+	 * @return the route node id
+	 */
+	public int getId() {
 		return 0;
 	}
