+++ src/uk/me/parabola/mkgmap/general/MapCollector.java
@@ -81,5 +81,5 @@ public interface MapCollector {
 	 * Add a through route to the map. 
 	 *
 	 */
-	public void addThroughRoute(long junctionNodeId, long roadIdA, long roadIdB);
+	public void addThroughRoute(int junctionNodeId, long roadIdA, long roadIdB);
 }
