+++ test/uk/me/parabola/mkgmap/osmstyle/StyledConverterTest.java
@@ -269,7 +269,7 @@ public class StyledConverterTest {
 				return 0;
 			}
 
-			public void addThroughRoute(long junctionNodeId, long roadIdA, long roadIdB) { }
+			public void addThroughRoute(int junctionNodeId, long roadIdA, long roadIdB) { }
 		};
 
 		return new StyledConverter(style, coll, new EnhancedProperties());
